+++ bms/assets/sql/1.0.0_to_1.0.1-RC1.sql
@@ -53,3 +53,34 @@ CREATE TABLE bdms.borehole_files
         ON DELETE NO ACTION
         NOT VALID
 );
+
+CREATE TABLE bdms.terms
+(
+    id_tes serial NOT NULL,
+    draft_tes boolean NOT NULL DEFAULT TRUE,
+    text_tes_en character varying NOT NULL,
+    text_tes_de character varying,
+    text_tes_fr character varying,
+    text_tes_it character varying,
+    text_tes_ro character varying,
+    creation_tes timestamp with time zone NOT NULL DEFAULT now(),
+    expired_tes timestamp with time zone,
+    PRIMARY KEY (id_tes)
+);
+
+CREATE TABLE bdms.terms_accepted (
+    id_usr_fk integer NOT NULL,
+    id_tes_fk integer NOT NULL,
+    accepted_tea timestamp with time zone DEFAULT now(),
+    PRIMARY KEY (id_usr_fk, id_tes_fk),
+    FOREIGN KEY (id_usr_fk)
+        REFERENCES bdms.users (id_usr) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE
+        NOT VALID,
+    FOREIGN KEY (id_tes_fk)
+        REFERENCES bdms.terms (id_tes) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE
+        NOT VALID
+);