+++ bms/v1/terms/__init__.py
@@ -1,0 +1,7 @@
+
+# Actions
+from bms.v1.terms.accept import AcceptTerms
+from bms.v1.terms.draft import DraftTerms
+from bms.v1.terms.get import GetTerms
+from bms.v1.terms.get import GetTermsDraft
+from bms.v1.terms.publish import PublishTerms